ist-publications-backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
